Bugfix + completed functions of task.service

+++ src/modules/project/project.service.ts
@@ -1,4 +1,8 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CreateProjectDTO } from './dto/create-project.dto';
 import { Project, ProjectDocument } from './model/project.model';
 import { Model } from 'mongoose';
@@ -27,18 +31,16 @@ export class ProjectService {
   }
 
   async createProject(
+    userId: string,
     data: CreateProjectDTO
   ): Promise<{ _id: string; name: string; imageUrl: string; teamId: string }> {
-    // check for authority to create project here
-
-    ///
     const { name, teamId, imageUrl } = data;
     const project = {
       name,
       teamId,
       imageUrl: imageUrl ? imageUrl : null,
+      ownerId: userId,
     };
-    // check for duplicate project generation in a team
     const teamProjects = await this.getProjectsByTeamId(teamId);
     if (teamProjects.some((project) => project.name === name)) {
       throw new ConflictException('Project already exists');
@@ -50,18 +52,26 @@ export class ProjectService {
     return { ...newProject };
   }
 
-  async removeProject(projectId: string) {
-    // check for authority to remove the project here
-
-    ///
+  async removeProject(userId: string, projectId: string) {
+    const project = await this.getProject(projectId);
+    if (!project) {
+      throw new ConflictException('Project not found');
+    }
+    if (project.ownerId !== userId) {
+      throw new UnauthorizedException('You are not the owner of this project');
+    }
     await this.projectModel.deleteOne({ _id: projectId });
     return true;
   }
 
-  async updateProject(projectId: string, query: object) {
-    // check for authority to update the project here
-
-    ///
+  async updateProject(userId: string, projectId: string, query: object) {
+    const project = await this.getProject(projectId);
+    if (!project) {
+      throw new ConflictException('Project not found');
+    }
+    if (userId !== project.ownerId) {
+      throw new UnauthorizedException('You are not the owner of this project');
+    }
     if (!projectId) throw new ConflictException('Project not found');
     if (!query) throw new ConflictException('Query not found');
     await this.projectModel.updateOne({ _id: projectId }, { $set: query });
